+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Create code style guide document. UIDEXP-148.
 * Implement transformations saving. UIDEXP-73.
 * Cover closing functionality of mapping profile form by tests. UIDEXP-99.
+* Implement edit mapping profile feature. UIDEXP-132.
 
 ## [2.0.1](https://github.com/folio-org/ui-data-export/tree/v2.0.1) (2020-07-09)
 [Full Changelog](https://github.com/folio-org/ui-data-export/tree/v2.0.0...v2.0.1)